Confirm messages added

+++ src/components/course/setting.js
@@ -33,13 +33,17 @@ class MinTemplate extends Component {
   }
 
   onDeleteInstance(instance) {
-    return instanceService.remove(instance.id)
-      .then(() => this.fetchInstances());
+    if (window.confirm("Do you really want to delete this course instance?")) {
+      return instanceService.remove(instance.id)
+        .then(() => this.fetchInstances());
+    }
   }
 
   onDeleteCourse() {
-    return courseService.remove(this.props.course.id)
-      .then(() => this.props.router.push(`/organizations/show/${this.props.course.organizationId}/courses`));
+    if (window.confirm("Do you really want to delete the entire course? ")) {
+      return courseService.remove(this.props.course.id)
+        .then(() => this.props.router.push(`/organizations/show/${this.props.course.organizationId}/courses`));
+    }
   }
 
   fetchInstances() {
